requirements.txt
blinker
Click
dnspython
Flask
Flask-Cors
Flask-Mail
Flask-Script
Flask-SQLAlchemy
Flask-WTF
idna
itsdangerous
Jinja2
MarkupSafe
PyMySQL
python-dateutil
requests
simplejson
six
SQLAlchemy
urllib3
Werkzeug
WTForms
